tg/requirements.txt
aiohttp
aiogram
beautifulsoup4
aiocache
python-dotenv
актуальные версии